# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Rest API",
      "version": "1.0.0",
      "title": "Rest API"
    },
    "schemes": [
      "http"
    ],
    "paths": {
      "/events": {
        "get": {
          "tags": [
            "events"
          ],
          "summary": "Get all avaliable events",
          "responses": {
            "200": {
              "description": "successful operation"
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
				},
				"put": {
          "tags": [
            "events"
          ],
          "summary": "Update event by ID",
          "description": "Updates a single event",
          "operationId": "updateEventById",
          "parameters": [
						{
							"in": "body",
							"name": "body",
							"description": "Updated object",
							"required": true,
							"schema": {
								"$ref": "#/definitions/Event"
							}
						}
					],
          "responses": {
            "200": {
							"description": "successful operation",
							"schema": {
								"$ref": "#/definitions/Event"
							}
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
				},
				"post": {
          "tags": [
            "events"
          ],
          "summary": "Create event",
          "description": "Create a single event",
          "operationId": "createEvent",
          "parameters": [
						{
							"in": "body",
							"name": "body",
							"description": "New event object",
							"required": true,
							"schema": {
								"$ref": "#/definitions/NewEvent"
							}
						}
					],
          "responses": {
            "200": {
							"description": "successful operation",
							"schema": {
								"$ref": "#/definitions/Event"
							}
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
        }
      },
      "/events/{eventId}": {
        "get": {
          "tags": [
            "events"
          ],
          "summary": "Find event by ID",
          "description": "Returns a single event",
          "operationId": "getEventById",
          "parameters": [
            {
              "name": "eventId",
              "in": "path",
              "description": "ID of event to return",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
							"description": "successful operation",
							"schema": {
								"$ref": "#/definitions/Event"
							}
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
				},
				"delete": {
          "tags": [
            "events"
          ],
          "summary": "Delete event by ID",
          "description": "Deletes a single event",
          "operationId": "deleteEventById",
          "parameters": [
            {
              "name": "eventId",
              "in": "path",
              "description": "ID of event to delete",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
							"description": "successful operation"
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
				}
      },
      "/discoveries": {
        "get": {
          "tags": [
            "discoveries"
          ],
          "summary": "Get all avaliable discoveries",
          "responses": {
            "200": {
              "description": "successful operation"
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
        },
        "put": {
          "tags": [
            "discoveries"
          ],
          "summary": "Update discovery by ID",
          "description": "Updates a single discovery",
          "operationId": "updateDiscoveryById",
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Updated object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Discovery"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Event"
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
        },
        "post": {
          "tags": [
            "discoveries"
          ],
          "summary": "Create discovery",
          "description": "Create a single discovery",
          "operationId": "createEvent",
          "parameters": [
						{
							"in": "body",
							"name": "body",
							"description": "New discovery object",
							"required": true,
							"schema": {
								"$ref": "#/definitions/NewDiscovery"
							}
						}
					],
          "responses": {
            "200": {
							"description": "successful operation",
							"schema": {
								"$ref": "#/definitions/Discovery"
							}
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Discovery not found"
            }
          }
        }
      },
      "/discoveries/{discoveryId}": {
        "get": {
          "tags": [
            "discoveries"
          ],
          "summary": "Find discovery by ID",
          "description": "Returns a single discovery",
          "operationId": "getDiscoveryById",
          "parameters": [
            {
              "name": "discoveryId",
              "in": "path",
              "description": "ID of discovery to return",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
							"description": "successful operation",
							"schema": {
								"$ref": "#/definitions/Discovery"
							}
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
        },
        "delete": {
          "tags": [
            "discoveries"
          ],
          "summary": "Delete discovery by ID",
          "description": "Deletes a single discovery",
          "operationId": "deleteDiscoveryById",
          "parameters": [
            {
              "name": "discoveryId",
              "in": "path",
              "description": "ID of discovery to delete",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
							"description": "successful operation"
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Event not found"
            }
          }
				}
      }
    },
    "definitions": {
        "Event": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "event": {
              "type": "string",
              "description": "Event"
            }
          },
          "xml": {
            "name": "Event"
          }
				},
				"NewEvent": {
          "type": "object",
          "properties": {
            "event": {
              "type": "string",
              "description": "Event"
            }
          },
          "xml": {
            "name": "NewEvent"
          }
        },
        "Discovery": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "city": {
              "type": "string",
              "description": "Discovery city"
            },
            "venue": {
              "type": "string",
              "description": "Discovery venue"
            }
          },
          "xml": {
            "name": "Event"
          }
        },
        "NewDiscovery": {
          "type": "object",
          "properties": {
            "city": {
              "type": "string",
              "description": "Discovery city"
            },
            "venue": {
              "type": "string",
              "description": "Discovery venue"
            }
          },
          "xml": {
            "name": "NewEvent"
          }
        }
    }
}